Move AgentPerson translation to separate module
